+++ python/PyQt6/core/auto_additions/qgslinechartplot.py
@@ -1,0 +1,7 @@
+# The following has been generated automatically from src/core/plot/qgslinechartplot.h
+try:
+    QgsLineChartPlot.create = staticmethod(QgsLineChartPlot.create)
+    QgsLineChartPlot.__overridden_methods__ = ['type', 'renderContent', 'writeXml', 'readXml']
+    QgsLineChartPlot.__group__ = ['plot']
+except (NameError, AttributeError):
+    pass
