+++ python/core/auto_additions/qgsbarchartplot.py
@@ -1,0 +1,7 @@
+# The following has been generated automatically from src/core/plot/qgsbarchartplot.h
+try:
+    QgsBarChartPlot.create = staticmethod(QgsBarChartPlot.create)
+    QgsBarChartPlot.__overridden_methods__ = ['type', 'renderContent', 'writeXml', 'readXml']
+    QgsBarChartPlot.__group__ = ['plot']
+except (NameError, AttributeError):
+    pass
