--- python/core/auto_additions/qgschartplot.py
@@ -1,13 +1,0 @@
-# The following has been generated automatically from src/core/plot/qgschartplot.h
-try:
-    QgsBarChartPlot.create = staticmethod(QgsBarChartPlot.create)
-    QgsBarChartPlot.__overridden_methods__ = ['type', 'renderContent', 'writeXml', 'readXml']
-    QgsBarChartPlot.__group__ = ['plot']
-except (NameError, AttributeError):
-    pass
-try:
-    QgsLineChartPlot.create = staticmethod(QgsLineChartPlot.create)
-    QgsLineChartPlot.__overridden_methods__ = ['type', 'renderContent', 'writeXml', 'readXml']
-    QgsLineChartPlot.__group__ = ['plot']
-except (NameError, AttributeError):
-    pass
