+++ python/core/auto_generated/plot/qgsbarchartplot.sip.in
@@ -1,7 +1,7 @@
 /************************************************************************
  * This file has been generated automatically from                      *
  *                                                                      *
- * src/core/plot/qgschartplot.h                                         *
+ * src/core/plot/qgsbarchartplot.h                                      *
  *                                                                      *
  * Do not edit manually ! Edit header and run scripts/sipify.py again   *
  ************************************************************************/
@@ -22,7 +22,7 @@ A simple bar chart class.
 %End
 
 %TypeHeaderCode
-#include "qgschartplot.h"
+#include "qgsbarchartplot.h"
 %End
   public:
 
@@ -56,68 +56,10 @@ Returns a new bar chart.
 
 };
 
-
-class QgsLineChartPlot : Qgs2DXyPlot
-{
-%Docstring(signature="appended")
-A simple line chart class.
-
-.. warning::
-
-   This class is not considered stable API, and may change in future!
-
-.. versionadded:: 4.0
-%End
-
-%TypeHeaderCode
-#include "qgschartplot.h"
-%End
-  public:
-
-    QgsLineChartPlot();
-    ~QgsLineChartPlot();
-
-    virtual QString type() const;
-
-    virtual void renderContent( QgsRenderContext &context, const QRectF &plotArea, const QgsPlotData &plotData = QgsPlotData() );
-
-
-    virtual bool writeXml( QDomElement &element, QDomDocument &document, const QgsReadWriteContext &context ) const;
-
-    virtual bool readXml( const QDomElement &element, const QgsReadWriteContext &context );
-
-
-    QgsMarkerSymbol *markerSymbol( int index ) const;
-%Docstring
-Returns the marker symbol at a given ``index``.
-%End
-
-    void setMarkerSymbol( int index, QgsMarkerSymbol *symbol /Transfer/ );
-%Docstring
-Sets the marker ``symbol`` at a given ``index``.
-%End
-
-    QgsLineSymbol *lineSymbol( int index ) const;
-%Docstring
-Returns the line symbol at a given ``index``.
-%End
-
-    void setLineSymbol( int index, QgsLineSymbol *symbol /Transfer/ );
-%Docstring
-Sets the marker ``symbol`` at a given ``index``.
-%End
-
-    static QgsLineChartPlot *create();
-%Docstring
-Returns a new line chart.
-%End
-
-};
-
 /************************************************************************
  * This file has been generated automatically from                      *
  *                                                                      *
- * src/core/plot/qgschartplot.h                                         *
+ * src/core/plot/qgsbarchartplot.h                                      *
  *                                                                      *
  * Do not edit manually ! Edit header and run scripts/sipify.py again   *
  ************************************************************************/
