+++ python/core/core_auto.sip
@@ -515,7 +515,8 @@
 %Include auto_generated/painting/qgspaintenginehack.sip
 %Include auto_generated/painting/qgspainting.sip
 %Include auto_generated/pdf/qgspdfrenderer.sip
-%Include auto_generated/plot/qgschartplot.sip
+%Include auto_generated/plot/qgsbarchartplot.sip
+%Include auto_generated/plot/qgslinechartplot.sip
 %Include auto_generated/plot/qgsplotregistry.sip
 %Include auto_generated/plot/qgsplot.sip
 %Include auto_generated/pointcloud/qgspointcloudattribute.sip
