+++ src/core/plot/qgslinechartplot.cpp
@@ -1,6 +1,6 @@
 /***************************************************************************
-                         qgschartplot.cpp
-                         ---------------
+                         qgslinechartplot.cpp
+                         --------------------
     begin                : June 2025
     copyright            : (C) 2025 by Mathieu
     email                : mathieu at opengis dot ch
@@ -15,174 +15,12 @@
  *                                                                         *
  ***************************************************************************/
 
-#include "qgschartplot.h"
+#include "qgslinechartplot.h"
 #include "qgsexpressioncontextutils.h"
 #include "qgssymbol.h"
 #include "qgssymbollayer.h"
 #include "qgssymbollayerutils.h"
-#include "qgsfillsymbol.h"
-#include "qgslinesymbol.h"
-#include "qgsmarkersymbol.h"
 
-void QgsBarChartPlot::renderContent( QgsRenderContext &context, const QRectF &plotArea, const QgsPlotData &plotData )
-{
-  if ( mFillSymbols.empty() )
-  {
-    return;
-  }
-
-  const QList<QgsAbstractPlotSeries *> seriesList = plotData.series();
-  if ( seriesList.isEmpty() )
-  {
-    return;
-  }
-
-  const QStringList categories = plotData.categories();
-  if ( xAxis().type() == Qgis::PlotAxisType::CategoryType && categories.isEmpty() )
-  {
-    return;
-  }
-
-  QgsExpressionContextScope *chartScope = new QgsExpressionContextScope( QStringLiteral( "chart" ) );
-  const QgsExpressionContextScopePopper scopePopper( context.expressionContext(), chartScope );
-
-  context.painter()->save();
-  context.painter()->setClipRect( plotArea );
-
-  const double xScale = plotArea.width() / ( xMaximum() - xMinimum() );
-  const double yScale = plotArea.height() / ( yMaximum() - yMinimum() );
-  const double categoriesWidth = plotArea.width() / categories.size();
-  const double barsWidth = categoriesWidth / 2;
-  const double barWidth = barsWidth / seriesList.size();
-  int seriesIndex = 0;
-  for ( const QgsAbstractPlotSeries *series : seriesList )
-  {
-    QgsFillSymbol *symbol = fillSymbol( seriesIndex % mFillSymbols.size() );
-    if ( !symbol )
-    {
-      continue;
-    }
-    symbol->startRender( context );
-
-    const double barStartAdjustement = -( barsWidth / 2 ) + barWidth * seriesIndex;
-    if ( const QgsXyPlotSeries *xySeries = dynamic_cast<const QgsXyPlotSeries *>( series ) )
-    {
-      const QList<std::pair<double, double>> data = xySeries->data();
-      for ( const std::pair<double, double> &pair : data )
-      {
-        if ( xAxis().type() == Qgis::PlotAxisType::CategoryType && ( pair.first < 0 || pair.first >= categories.size() ) )
-        {
-          continue;
-        }
-
-        double x = 0;
-        if ( xAxis().type() == Qgis::PlotAxisType::ValueType )
-        {
-          x = ( pair.first - xMinimum() ) * xScale + barStartAdjustement;
-        }
-        else if ( xAxis().type() == Qgis::PlotAxisType::CategoryType )
-        {
-          x = ( categoriesWidth * pair.first ) + ( categoriesWidth / 2 ) + barStartAdjustement;
-        }
-        double y = ( pair.second - yMinimum() ) * yScale;
-
-        const double zero = ( 0.0 - yMinimum() ) * yScale;
-        const QPoint topLeft( plotArea.left() + x,
-                              plotArea.y() + plotArea.height() - y );
-        const QPoint bottomRight( plotArea.left() + x + barWidth,
-                                  plotArea.y() + plotArea.height() - zero );
-
-        chartScope->addVariable( QgsExpressionContextScope::StaticVariable( QStringLiteral( "chart_value" ), pair.second, true ) );
-        symbol->renderPolygon( QPolygonF( QRectF( topLeft, bottomRight ) ), nullptr, nullptr, context );
-      }
-    }
-
-    symbol->stopRender( context );
-    seriesIndex++;
-  }
-
-  context.painter()->restore();
-}
-
-QgsFillSymbol *QgsBarChartPlot::fillSymbol( int index ) const
-{
-  if ( index < 0 || index >= static_cast<int>( mFillSymbols.size() ) )
-  {
-    return nullptr;
-  }
-
-  return mFillSymbols[index].get();
-}
-
-void QgsBarChartPlot::setFillSymbol( int index, QgsFillSymbol *symbol )
-{
-  if ( index < 0 )
-  {
-    return;
-  }
-
-  if ( index + 1 >= static_cast<int>( mFillSymbols.size() ) )
-  {
-    mFillSymbols.resize( index + 1 );
-  }
-
-  mFillSymbols[index].reset( symbol );
-}
-
-bool QgsBarChartPlot::writeXml( QDomElement &element, QDomDocument &document, const QgsReadWriteContext &context ) const
-{
-  Qgs2DXyPlot::writeXml( element, document, context );
-
-  QDomElement fillSymbolsElement = document.createElement( QStringLiteral( "fillSymbols" ) );
-  for ( int i = 0; i < static_cast<int>( mFillSymbols.size() ); i++ )
-  {
-    QDomElement fillSymbolElement = document.createElement( QStringLiteral( "fillSymbol" ) );
-    fillSymbolElement.setAttribute( QStringLiteral( "index" ), QString::number( i ) );
-    if ( mFillSymbols[i] )
-    {
-      fillSymbolElement.appendChild( QgsSymbolLayerUtils::saveSymbol( QString(), mFillSymbols[i].get(), document, context ) );
-    }
-    fillSymbolsElement.appendChild( fillSymbolElement );
-  }
-  element.appendChild( fillSymbolsElement );
-
-  return true;
-}
-
-bool QgsBarChartPlot::readXml( const QDomElement &element, const QgsReadWriteContext &context )
-{
-  Qgs2DXyPlot::readXml( element, context );
-
-  const QDomNodeList fillSymbolsList = element.firstChildElement( QStringLiteral( "fillSymbols" ) ).childNodes();
-  for ( int i = 0; i < fillSymbolsList.count(); i++ )
-  {
-    const QDomElement fillSymbolElement = fillSymbolsList.at( i ).toElement();
-    const int index = fillSymbolElement.attribute( QStringLiteral( "index" ), QStringLiteral( "-1" ) ).toInt();
-    if ( index >= 0 )
-    {
-      if ( fillSymbolElement.hasChildNodes() )
-      {
-        const QDomElement symbolElement = fillSymbolElement.firstChildElement( QStringLiteral( "symbol" ) );
-        setFillSymbol( index, QgsSymbolLayerUtils::loadSymbol< QgsFillSymbol >( symbolElement, context ).release() );
-      }
-      else
-      {
-        setFillSymbol( index, nullptr );
-      }
-    }
-  }
-
-  return true;
-}
-
-QgsBarChartPlot *QgsBarChartPlot::create()
-{
-  return new QgsBarChartPlot();
-}
-
-//
-// QgsLineChartPlot
-//
 
 void QgsLineChartPlot::renderContent( QgsRenderContext &context, const QRectF &plotArea, const QgsPlotData &plotData )
 {
